feat: implement double-checked locking for singleton analytics instance
